+++ hardpy/pytest_hardpy/utils/node_info.py
@@ -47,9 +47,8 @@ class NodeInfo:
 
         self._critical = self._get_critical(item.own_markers + item.parent.own_markers)
 
-        # TODO(xorialexandrov):  Add module and case group
-        self._module_group = Group.MAIN
-        self._case_group = Group.MAIN
+        self._module_group = self._get_group(item.parent.own_markers, "module_group")
+        self._case_group = self._get_group(item.own_markers, "case_group")
 
         self._module_id = Path(item.parent.nodeid).stem  # type: ignore
         self._case_id = item.name
@@ -222,3 +221,33 @@ class NodeInfo:
             bool: True if test or module is critical, False otherwise
         """
         return any(marker.name == "critical" for marker in markers)
+
+    def _get_group(
+        self,
+        markers: list[Mark],
+        marker_name: str,
+    ) -> Group:
+        """Get group from markers or use default.
+
+        Args:
+            markers (list[Mark]): item markers list
+            marker_name (str): marker name
+        Returns:
+            Group: group from marker or default (Group.MAIN)
+        """
+        for marker in markers:
+            if marker.name == marker_name and marker.args:
+                arg = marker.args[0]
+
+                if isinstance(arg, Group):
+                    return arg
+                if isinstance(arg, str):
+                    valid_groups = {group.value for group in Group}
+                    if arg not in valid_groups:
+                        msg = f"Invalid group '{arg}'. Valid groups are: {', '.join(valid_groups)}"  # noqa: E501
+                        raise ValueError(msg)
+                    return Group(arg)
+                msg = f"Group marker argument must be either string or Group enum, got {type(arg)}"  # noqa: E501
+                raise ValueError(msg)
+
+        return Group.MAIN